Fixing logger handling to be applied by workers
