+++ biocentral_server/predict/multi_prediction_task.py
@@ -4,7 +4,10 @@ from biotrainer.utilities import get_device
 from .models import BaseModel
 from .single_prediction_task import SinglePredictionTask
 
+from ..utils import get_logger
 from ..server_management import TaskInterface, TaskDTO
+
+logger = get_logger(__name__)
 
 
 class MultiPredictionTask(TaskInterface):
@@ -27,7 +30,7 @@ class MultiPredictionTask(TaskInterface):
             if not predict_dto:
                 return TaskDTO.failed(error=f"Model prediction with the {model_name} model failed.")
             single_prediction = predict_dto.update["predictions"]
-            print(f"{model_name} model prediction: {single_prediction}")
+            logger.info(f"{model_name} model prediction: {single_prediction}")
             predictions[model_name] = single_prediction
 
         return TaskDTO.finished(result={"predictions": predictions})
