+++ biocentral_server/utils/__init__.py
@@ -1,7 +1,9 @@
 from .format_utils import str2bool
 from .constants import Constants
+from .logging import get_logger
 
 __all__ = [
     'str2bool',
     'Constants',
+    'get_logger'
 ]
